Move test to `test_resources.py`
